Moved the GroupListAdapater to its own file
